test: move e2e tests to e2e directory
